+++ example/lib/views/widgets/file_type_icon.dart
@@ -24,9 +24,6 @@ class FileTypeIcon extends StatelessWidget {
             if (input.path != null) {
               return FileMagicNumber.detectFileTypeFromPathOrBlob(input.path!);
             }
-            if (input.url != null) {
-              return FileMagicNumber.detectFileTypeFromPathOrBlob(input.url!);
-            }
             return Future.value(null);
           }(),
           builder: (context, snapshot) {